Add UnitOfWork pattern implementation with transaction management and repositories
Implemented the UnitOfWork class to centralize the management of database transactions.

Added repositories for User, RefreshToken, Category, and Product entities using the Repository pattern.

Introduced transaction management methods:

BeginTransactionAsync to initiate a database transaction.

CommitTransactionAsync to commit the transaction if no errors occur.

RollbackTransactionAsync to rollback the transaction in case of an error.

Ensured proper disposal of resources by implementing the Dispose method to clean up transactions and the database context.

Provides a consistent and efficient way of managing database operations and transactions across the application.